simplified twitter embed/content
no point in putting the tweet in a blockquote or stating the twitterers
name (it /me) or stating the date or the backlink. it’s all in the
post’s meta data:
author, link, date
basically the tweet is striped down to the bare content. which makes
sense for one’s own tweets.